Move protocol files to versioned folders
