+++ tests/unit/test_protocol_json/test_basic_simple_protocol_json.py
@@ -1,8 +1,6 @@
 """
 Test JSON creation for basic simple protocol.
 """
-import pytest
-from model_train_protocol import Protocol
 from tests.utils.protocol_json_utils import assert_special_tokens_in_tokens
 
 
@@ -12,7 +10,7 @@ class TestBasicSimpleProtocolJSON:
     def _get_json_output(self, protocol):
         """Helper method to get JSON output from a protocol."""
         protocol._prep_protocol()
-        from model_train_protocol.integration.ProtocolFile import ProtocolFile
+        from model_train_protocol.versioning.files.protocol_file.v1.ProtocolFile import ProtocolFile
         protocol_file = ProtocolFile(
             name=protocol.name,
             context=protocol.context,
