+++ tests/unit/test_protocol_json/test_numlisttoken_protocol_json.py
@@ -1,8 +1,6 @@
 """
 Test JSON creation for NumListToken protocol.
 """
-import pytest
-from model_train_protocol import Protocol
 
 
 class TestNumListTokenProtocolJSON:
@@ -11,7 +9,7 @@ class TestNumListTokenProtocolJSON:
     def _get_json_output(self, protocol):
         """Helper method to get JSON output from a protocol."""
         protocol._prep_protocol()
-        from model_train_protocol.integration.ProtocolFile import ProtocolFile
+        from model_train_protocol.versioning.files.protocol_file.v1.ProtocolFile import ProtocolFile
         protocol_file = ProtocolFile(
             name=protocol.name,
             context=protocol.context,
